Add print and print_summary to writers.
