+++ gcovr/writer/csv/__init__.py
@@ -20,7 +20,7 @@
 import csv
 from typing import List, Tuple, Optional
 
-from ...options import GcovrConfigOption, OutputOrDefault
+from ...options import GcovrConfigOption, Options, OutputOrDefault
 from ...writer.base import writer_base
 
 from ...utils import presentable_filename, open_text_for_writing
@@ -46,49 +46,7 @@ class writer(writer_base):
             ),
         ]
 
+    def print_report(covdata: CovData, output_file: str, options: Options) -> bool:
+        from .print import print_report
 
-def print_csv_report(covdata: CovData, output_file, options):
-    """produce gcovr csv report"""
-
-    with open_text_for_writing(output_file, "coverage.csv", newline="") as fh:
-        keys = sort_coverage(
-            covdata,
-            show_branch=options.show_branch,
-            by_num_uncovered=options.sort_uncovered,
-            by_percent_uncovered=options.sort_percent,
-        )
-
-        writer = csv.writer(fh)
-        writer.writerow(
-            (
-                "filename",
-                "line_total",
-                "line_covered",
-                "line_percent",
-                "branch_total",
-                "branch_covered",
-                "branch_percent",
-                "function_total",
-                "function_covered",
-                "function_percent",
-            )
-        )
-        for key in keys:
-            filename = presentable_filename(covdata[key].filename, options.root_filter)
-            stats = SummarizedStats.from_file(covdata[key])
-            writer.writerow(
-                [
-                    filename,
-                    *_stat_tuple(stats.line),
-                    *_stat_tuple(stats.branch),
-                    *_stat_tuple(stats.function),
-                ]
-            )
-
-
-def _stat_tuple(stat: CoverageStat) -> Tuple[int, int, Optional[float]]:
-    """creates tuple (total, covered, ratio) with ratio in range 0..1 incl"""
-    percent = stat.percent
-    if percent is not None:
-        percent = percent / 100.0
-    return stat.total, stat.covered, percent
+        return print_report(covdata, output_file, options)
