+++ gcovr/writer/sonarqube/__init__.py
@@ -20,10 +20,9 @@
 from typing import List
 from lxml import etree
 
-from ...options import GcovrConfigOption, OutputOrDefault
+from ...options import GcovrConfigOption, Options, OutputOrDefault
 from ...writer.base import writer_base
 
-from ...utils import open_binary_for_writing, presentable_filename
 from ...coverage import CovData
 
 
@@ -46,41 +45,7 @@ class writer(writer_base):
             ),
         ]
 
+    def print_report(covdata: CovData, output_file: str, options: Options) -> bool:
+        from .print import print_report
 
-def print_sonarqube_report(covdata: CovData, output_file, options):
-    """produce an XML report in the Sonarqube generic coverage format"""
-
-    root = etree.Element("coverage")
-    root.set("version", "1")
-
-    for f in sorted(covdata):
-        data = covdata[f]
-        filename = presentable_filename(f, root_filter=options.root_filter)
-
-        fileNode = etree.Element("file")
-        fileNode.set("path", filename)
-
-        for lineno in sorted(data.lines):
-            line_cov = data.lines[lineno]
-            if not line_cov.is_covered and not line_cov.is_uncovered:
-                continue
-
-            L = etree.Element("lineToCover")
-            L.set("lineNumber", str(lineno))
-            if line_cov.is_covered:
-                L.set("covered", "true")
-            else:
-                L.set("covered", "false")
-
-            branches = line_cov.branches
-            if branches:
-                b = line_cov.branch_coverage()
-                L.set("branchesToCover", str(b.total))
-                L.set("coveredBranches", str(b.covered))
-
-            fileNode.append(L)
-
-        root.append(fileNode)
-
-    with open_binary_for_writing(output_file, "sonarqube.xml") as fh:
-        fh.write(etree.tostring(root, encoding="UTF-8", xml_declaration=True))
+        return print_report(covdata, output_file, options)
